requirements.txt

google-genai
aiohttp==3.11.11
